+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v0.3.0 (in development)
 -----------------------
 - Switch from nom to winnow
+- `release`: Take name of repository's default branch into account when adding
+  changelog link to README
 
 v0.2.0 (2023-11-22)
 -------------------